services/userservice/requirements.txt
Django>=4.2,<5.0
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0
whitenoise>=6.6,<7.0
